Moves the test data to the test directory and changes the test to call the function
